requirements.txt
numpy~=1.21.4
PyQt5~=5.15.6
pyqtgraph~=0.12.3
QtPy~=1.11.3
ijson~=3.1.4
pandas~=1.3.5
cdflib~=0.3.20
h5py~=3.6.0
tqdm~=4.62.3
torch~=1.10.1+cu113
torchvision~=0.11.2+cu113
matplotlib~=3.5.1
scipy~=1.7.3
packaging~=21.3
pycocotools~=2.0.3
PyOpenGL-accelerate~=3.1.5
opencv-python-headless~=4.5.4.60
scikit-learn~=1.0.1